Made input manager, WIP config manager, hpp -> h
  * Wrote an input manager from start to finish. Offers Unity-like
capabilities (GetButton("Action")) and includes Down and Up variants for
frame-specific input without tracking booleans for that. Currently
supports all keyboard and mouse input, joysticks will need some work to
get them in (have to handle all of the axes as inputs, fun). The current
implementation is not quite to my liking, but seeing as how SFML wants
to segregate inputs into different static namespace modules, this is
really the only way to pull something like this off. It's not that ugly, and
I'm not mad... but I am just disappointed.

  * WIP config system. I don't know yet if I want a static class with an
internal configuration stored privately, or if I want it to act like a static helper
class for dealing with Configuration structs that are kept track of in
the main game loop. May want to talk with Ittai and ask for his input on
it.

  * Renamed hpp files to .h files. The logic for this was that, well...
to be honest, I don't see many hpp files in the wild, .h files are just
fine and take up less characters. I don't know yet the benefits between
the two, just that I would like to adhere to C++isms, and most people
use .h... so we'll stick with that as well.